+++ source/benoit/benoit/render/render_data.rs
@@ -21,7 +21,7 @@
 	If not, see <https://www.gnu.org/licenses/>.
 */
 
-use crate::benoit::width_height_ratio;
+use crate::benoit::{PRECISION, width_height_ratio};
 
 extern crate rug;
 
@@ -32,11 +32,15 @@ pub struct RenderData {
 	canvas_width:  u32,
 	canvas_height: u32,
 
+	canvas_size: usize,
+
 	centre_real: Float,
 	centre_imag: Float,
 	zoom:        Float,
 
 	max_iter_count: u32,
+
+	inverse: bool,
 
 	x_offset: f32,
 	y_offset: f32,
@@ -50,18 +54,22 @@ pub struct RenderData {
 
 impl RenderData {
 	#[must_use]
-	pub fn new(iter_count_buffer: &mut [u32], square_dist_buffer: &mut [f32], canvas_width: u32, canvas_height: u32, centre_real: Float, centre_imag: Float, zoom: Float, max_iter_count: u32) -> RenderData {
+	pub fn new(iter_count_buffer: &mut [u32], square_dist_buffer: &mut [f32], canvas_width: u32, canvas_height: u32, centre_real: Float, centre_imag: Float, zoom: Float, max_iter_count: u32, inverse: bool) -> RenderData {
 		let (width_ratio, height_ratio) = width_height_ratio(canvas_width, canvas_height);
 
 		return RenderData {
 			canvas_width:  canvas_width,
 			canvas_height: canvas_height,
 
+			canvas_size: canvas_height as usize * canvas_width as usize,
+
 			centre_real: centre_real,
 			centre_imag: centre_imag,
 			zoom:        zoom,
 
 			max_iter_count: max_iter_count,
+
+			inverse: inverse,
 
 			x_offset: canvas_width  as f32 / -2.0,
 			y_offset: canvas_height as f32 / -2.0,
@@ -75,18 +83,32 @@ impl RenderData {
 	}
 
 	#[must_use]
-	pub fn input(&self) -> (u32, &Float, &Float, &Float, u32) {
-		return (self.canvas_width, &self.centre_real, &self.centre_imag, &self.zoom, self.max_iter_count);
+	pub fn inverse_factor(&self, a: &Float, b: &Float) -> Float {
+		return if self.inverse {
+			let mut inverse_factor = Float::with_val(PRECISION, a * a);
+			inverse_factor += b * b;
+			inverse_factor.recip_mut();
+
+			inverse_factor
+		} else {
+			Float::with_val(PRECISION, 0x1)
+		};
 	}
 
 	#[must_use]
-	pub fn output_buffers(&self, row: u32) -> (&mut [u32], &mut [f32]) {
-		assert!(row < self.canvas_height);
+	pub fn canvas_size(&self) -> (u32, u32) {
+		return (self.canvas_width, self.canvas_height);
+	}
 
-		let offset = row as isize * self.canvas_width as isize;
+	#[must_use]
+	pub fn input(&self) -> (&Float, &Float, &Float, u32) {
+		return (&self.centre_real, &self.centre_imag, &self.zoom, self.max_iter_count);
+	}
 
-		let iter_count  = unsafe { from_raw_parts_mut(self.iter_count_buffer.offset(offset),  self.canvas_width as usize) };
-		let square_dist = unsafe { from_raw_parts_mut(self.square_dist_buffer.offset(offset), self.canvas_width as usize) };
+	#[must_use]
+	pub fn output_buffers(&self) -> (&mut [u32], &mut [f32]) {
+		let iter_count  = unsafe { from_raw_parts_mut(self.iter_count_buffer,  self.canvas_size as usize) };
+		let square_dist = unsafe { from_raw_parts_mut(self.square_dist_buffer, self.canvas_size as usize) };
 
 		return (iter_count, square_dist);
 	}